feat: consolidate code for new CopyToClipboard component
