--- packages/clipboard-alpha/README.md
@@ -1,14 +1,0 @@
-# @launchpad-ui/clipboard
-
-> An element that copies text to the clipboard.
-
-[![See it on NPM!](https://img.shields.io/npm/v/@launchpad-ui/clipboard?style=for-the-badge)](https://www.npmjs.com/package/@launchpad-ui/clipboard)
-[![How big is this package in your project?](https://img.shields.io/bundlephobia/minzip/@launchpad-ui/clipboard?style=for-the-badge)](https://bundlephobia.com/result?p=@launchpad-ui/clipboard)
-
-## Installation
-
-```sh
-$ yarn add @launchpad-ui/clipboard
-# or
-$ npm install @launchpad-ui/clipboard
-```
