--- packages/clipboard-alpha/src/index.ts
@@ -1,5 +1,0 @@
-export type { CopyToClipboardProps, CopyToClipboardHandleRef } from './CopyToClipboard';
-export { CopyConfirmation, CopyToClipboard } from './CopyToClipboard';
-
-export type { CopyCodeButtonProps } from './CopyCodeButton';
-export { CopyCodeButton } from './CopyCodeButton';
