--- packages/clipboard-alpha/src/styles/CopyToClipboard.module.css
@@ -1,15 +1,0 @@
-.CopyToClipboard {
-  position: relative;
-}
-
-.Clipboard-checkmark svg {
-  fill: var(--lp-color-system-green-500);
-}
-
-.Clipboard-confirmation {
-  align-items: center;
-}
-
-.Clipboard-copied {
-  margin-left: 0.5rem;
-}
