--- packages/clipboard-alpha/tsconfig.build.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "include": ["src/*.ts", "src/*.tsx", "../../types/declarations.d.ts"],
-  "compilerOptions": {
-    "outDir": "./dist"
-  }
-}
